bugfix : merge 후 생긴 버그 수정
