+++ build.gradle
@@ -27,6 +27,7 @@ dependencies {
     implementation 'io.jsonwebtoken:jjwt:0.5.1'
     implementation 'org.modelmapper:modelmapper:2.4.4'
     implementation 'org.projectlombok:lombok:1.18.18'
+    implementation 'com.querydsl:querydsl-jpa'
     implementation group: 'io.springfox', name: 'springfox-boot-starter', version: '3.0.0'
     implementation group: 'com.google.guava', name: 'guava', version: '30.1.1-jre'
     implementation group: 'org.mariadb.jdbc', name: 'mariadb-java-client', version: '2.7.2'
